+++ client/src/components/profile/ModifyInfoCard/style.ts
@@ -1,0 +1,67 @@
+import styled, { css } from 'styled-components';
+
+export const Wrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  width: 45rem;
+`;
+
+export const InfoWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  margin: 3px;
+`;
+export const Legend = styled.legend`
+  font-size: 3rem;
+  color: ${({ theme }) => theme.colors.bgGreen};
+`;
+
+export const Info = styled.div`
+  font-size: 5rem;
+  color: ${({ theme }) => theme.colors.bgGreen};
+  width: 40rem;
+`;
+
+export const InfoSet = styled.fieldset`
+  ${({ theme }) => css`
+    background-color: ${theme.colors.grey};
+    padding: ${theme.paddings.lg};
+    border: 1px solid ${theme.colors.borderGrey};
+    border-radius: ${theme.borderRadius.base};
+  `};
+`;
+
+export const ModifyBtn = styled.span`
+  ${({ theme }) => theme.flexCenter}
+  width:100%;
+  background-color: ${({ theme }) => theme.colors.green};
+  padding: ${({ theme }) => theme.paddings.sm};
+  color: ${({ theme }) => theme.colors.white};
+  border-radius: 1rem;
+  cursor: pointer;
+  :hover {
+    color: ${({ theme }) => theme.colors.primary};
+  }
+`;
+
+export const CancelBtn = styled.span`
+  ${({ theme }) => theme.flexCenter}
+  width:100%;
+  background-color: ${({ theme }) => theme.colors.secondary};
+  padding: ${({ theme }) => theme.paddings.sm};
+  color: ${({ theme }) => theme.colors.white};
+  border-radius: 1rem;
+  cursor: pointer;
+  :hover {
+    color: ${({ theme }) => theme.colors.primary};
+  }
+`;
+
+export const Input = styled.input`
+  font-size: 5rem;
+`;
+
+export const ButtonWrapper = styled.div`
+  display: flex;
+  flex-direction: row;
+`;
